Move tracer auto-resolve module
